refactor(ui): share the panel and analysis primitives, and name the solar status by product
Preparation for the energy screen: everything that screen needs to render a
result already exists, but all of it was private to the page or the panel that
happened to be built first. A second surface could not show a result without
re-implementing what a result looks like.

Moved, with no behaviour change:

- The section heading was duplicated byte for byte in the classification,
  surface-water and solar panels. One definition now, with the step number
  optional -- that is what separates a sequence from a set. The first two number
  their sections because they are stages of one run; a panel of independently
  runnable products passes no step, because numbering them asserts an order that
  does not exist.
- The six presentation primitives, into components/analysisPrimitives.
- The energy model with its seven sub-components, and the wind screening, into
  their own files. AnalysisPage drops from 3892 to 2320 lines and imports them
  back, so there is one definition of each and the two screens cannot drift.

Three corrections to the solar status panel, all cases of it describing itself
wrongly:

Its heading was the fixed string "Solar resource" while the figures below it
could be a suitability classification or a plane-of-array raster -- a different
product with different units. It now names the product it is showing, in the
same words Overlay Tools uses for the same two rasters.

A point-resource summary was gated on no raster being present, so making a
terrain or siting map silently removed a result the user had just computed. The
block is additive now, and labelled when a raster shares the panel.

Its left edge was hardcoded to the classification panel's width, so it was
positioned by a panel it has nothing to do with. That is a prop.
