+++ frontend/src/components/SolarPanel.tsx
@@ -13,27 +13,8 @@ import {
   Zap,
 } from "lucide-react"
 import { cn } from "@/lib/utils"
+import { PanelSection as Section } from "@/components/ui/PanelSection"
 import type { SolarSeason } from "@/lib/types"
-
-function Section({
-  step,
-  title,
-  children,
-}: {
-  step: string
-  title: string
-  children: React.ReactNode
-}) {
-  return (
-    <div className="flex flex-col gap-2">
-      <div className="flex items-center gap-2">
-        <span className="telemetry text-[10px] text-primary">{step}</span>
-        <span className="eyebrow !text-foreground">{title}</span>
-      </div>
-      {children}
-    </div>
-  )
-}
 
 /**
  * What an action returns, stated before it is run.
